+++ CHANGELOG.md
@@ -24,6 +24,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - `Grax.Schema.schema?/1` and `Grax.Schema.struct?/1` to determine whether a given
   module or struct is a Grax schema resp. Grax schema struct
 - `Grax.schema/1` to get the schema(s) of a given class IRI
+- `Grax.Schema.schema?/1` to check if a given module or struct is a `Grax.Schema`
+- `Grax.Schema.inherited_from?/1` to check if a given module or struct is 
+  inherited from another `Grax.Schema`
 - `Grax.delete_additional_predicates/2` to delete all additional statements
 
 ### Changed